backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
requests
pytest
python-dotenv
openai
anthropic
google-generativeai